move user and csv logic to utility files
