+++ chirps.csv
@@ -2,3 +2,7 @@
 18965,joekane,whats up,public
 18362,aetherly,hey i'm alan,public
 19134,rtanay,hi i'm ryan,public
+15780,aetherly,does this still work?,public
+17330,sbrown,what about this?,public
+13083,ekane,hi i'm emily,public
+12623,aetherly,yo this is alan i'm back,public
