+++ main.py
@@ -1,9 +1,10 @@
 import time
 from user import *
-from csv_methods import *
+from csv_utility import *
 from random_id_generator import *
 from chirps import *
 from view_chirps import *
+from user_utility import *
 
 class MainMenu:
 
@@ -13,8 +14,8 @@ class MainMenu:
 		self.users = dict()
 		self.current_user = None
 		try:
-			self.users = get_users_from_csv_file("users.csv")
-			self.chirps = get_chirps_from_csv_file("chirps.csv")
+			self.users = CSV.get_users_from_csv_file("users.csv")
+			self.chirps = CSV.get_chirps_from_csv_file("chirps.csv")
 		except FileNotFoundError:
 			pass
 
@@ -37,38 +38,19 @@ class MainMenu:
 		try:
 			if int(action) > 0 and int(action) < 7:
 				if action == "1":
-					print("")
-					print("Enter full name")
-					full_name = input("> ")
-					print("")
-					print("Enter screen name")
-					screen_name = input("> ")
-					user = User(full_name, screen_name, random_id_generator())
-					write_user_to_csv_file(user, "users.csv")
+					user = UserUtility.create_user()
+					CSV.write_user_to_csv_file(user, "users.csv")
 					self.current_user = user
+					self.users = CSV.get_users_from_csv_file("users.csv")
 					print("Welcome, {}!".format(self.current_user.screen_name))
 
 				elif action == "2":
-					print("")
-					print("Please choose a user:")
-					counter = 1
-					user_list = list()
-					for k, v in self.users.items():
-						print("{0}: {1}".format(counter, v[0]))
-						user_list.append(k)
-						counter += 1
-					selection = input("> ")
-					selected_user_id = user_list[int(selection) - 1]
-					current_user = User(
-						self.users[selected_user_id][1],
-						self.users[selected_user_id][0],
-						selected_user_id
-					)
-					self.current_user = current_user
+					self.current_user = UserUtility.select_a_user(self.users)
 					print("Welcome, {}!".format(self.current_user.screen_name))
 
 				elif action == "3":
 					view_chirps(self.chirps)
+
 				elif action == "4":
 					if self.current_user:
 						print("")
@@ -79,7 +61,7 @@ class MainMenu:
 							self.current_user.screen_name,
 							random_id_generator()
 						)
-						write_public_chirp_to_csv_file(public_chirp, "chirps.csv")
+						CSV.write_public_chirp_to_csv_file(public_chirp, "chirps.csv")
 
 					else:
 						print("")
