+++ test_csv_methods.py
@@ -1,6 +1,6 @@
 import unittest
 from user import *
-from csv_methods import *
+from csv_utility import *
 from random_id_generator import *
 
 class TestCSVMethods(unittest.TestCase):
@@ -8,8 +8,8 @@ class TestCSVMethods(unittest.TestCase):
 	def test_csv_methods_return_user_dict(self):
 		test_user_dict = {19108: ["jkane", "joe"]}
 		user = User("joe", "jkane", random_id_generator())
-		write_user_to_csv_file(user, "test_csv.csv")
-		user_dict = get_users_from_csv_file("test_csv.csv")
+		CSV.write_user_to_csv_file(user, "test_csv.csv")
+		user_dict = CSV.get_users_from_csv_file("test_csv.csv")
 		for key, value in user_dict.items():
 			self.assertEqual(value, test_user_dict[19108])
